// src/api/model/apiResponseHostProfileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */
import type { ApiResponseHostProfileResponseCode } from './apiResponseHostProfileResponseCode';
import type { HostProfileResponse } from './hostProfileResponse';

export interface ApiResponseHostProfileResponse {
  code?: ApiResponseHostProfileResponseCode;
  message?: string;
  data?: HostProfileResponse;
}